requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
-e git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33#egg=clip
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.45.1
fsspec==2023.10.0
ftfy==6.1.3
idna==3.6
imageio==2.33.0
importlib-resources==6.1.1
Jinja2==3.1.2
kiwisolver==1.4.5
lazy-loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.4
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
Pillow==10.1.0
pycocotools==2.0.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-fid==0.3.0
PyWavelets==1.4.1
regex==2023.10.3
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
sympy==1.12
tifffile==2023.7.10
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
triton==2.1.0
typing-extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.12
zipp==3.17.0
